% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphantasiaWMStrats-package.R
\docType{package}
\name{aphantasiaWMStrats-package}
\alias{aphantasiaWMStrats}
\alias{aphantasiaWMStrats-package}
\title{aphantasiaWMStrats: Data Analysis for a Working Memory Strategies Study on Aphantasia}
\description{
This package contains the data and code to reproduce the analyses associated with an online study on a complex working memory task investigating non-visual working memory strategies in aphantasia.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://m-delem.github.io/aphantasiaWMStrats}
  \item \url{https://m-delem.github.io/aphantasiaWMStrats}
}

}
\author{
\strong{Maintainer}: Maël Delem \email{mael.delem@pm.me} (\href{https://orcid.org/0009-0005-8518-1991}{ORCID})

Authors:
\itemize{
  \item Maël Delem \email{mael.delem@pm.me} (\href{https://orcid.org/0009-0005-8518-1991}{ORCID})
}

}
\keyword{internal}
