% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tools.R
\name{scale_x_vviq}
\alias{scale_x_vviq}
\title{Custom x-axis scale for VVIQ scores}
\usage{
scale_x_vviq(
  breaks = seq(16, 80, by = 8),
  expand = ggplot2::expansion(mult = c(0.02, 0.02)),
  ...
)
}
\arguments{
\item{breaks}{Breaks for the x-axis. Default is \code{seq(16, 80, by = 8)}.}

\item{expand}{Expansion for the x-axis.}

\item{...}{Additional arguments passed to \code{ggplot2::scale_x_continuous()}.}
}
\value{
A list containing a ggplot2 scale object for the x-axis.
}
\description{
Ported from \code{aphantasiaEmotions}.
}
